requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
MetaTrader5>=5.0.4500
ccxt>=4.0.0
oandapyV20>=0.7.0
python-dotenv>=1.0.0
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
ta>=0.11.0
requests>=2.31.0
pydantic>=2.5.0
loguru>=0.7.0
